Split polar ECharts contract tests
